Fix bug where imports would fail based on the time of day.
This bug was caused by the discrepancy between the server date (UTC) and
the local time (inputs from the client).
When the transaction builder was called in the import process, the server
date was one day behind the current local date, causing transactions for
the current local date being rejected as "future dates".

Changes:
- Add CLI argument to specify timezone and add timezone to app state.
- Use local time when getting current time.
- Refactor interface for `Transaction` and `TransactionBuilder`.
  Removes most of the accessor functions and shifts
  validation of dates to `TransactionBuilder::finalize(..)`.