chore: Make release workflow manually triggered
This will allow multiple PRs to be merged into main and then released at
once. Currently, releases with multiple changes require multiple PRs to
be merged into a feature branch off of main and then merged via a single
PR into main. This led to duplicate commit messages and required me to
come up with a commit message that summarised all the PRs. This change
will allow for more control over the release process as well as allow me
to relax the rules for merging into main.