feat: Redirect back to same page in paginated transactions after editing
The current behaviour when editing a transaction is to redirect to the
transaction page, but the current page is not restored. So if you were on
page 10 and edit a transaction, you will be sent back to page 1.
The new behaviour will take you back to the page you started on and also
the redirect will trigger a page load which will show the updated
transaction.