chore: Relax unique versioning requirements
Update release process so that PR into main only requires a version bump
if the version in the local cargo.toml exists as a tag in the container
registry. This will ensure that the version number is bumped for the next
release while allowing refactor and chore type commits to be merged
without unnecessarily bumping the version.