feat: Autofocus inputs in forms (#66)
Autofocus all new item forms and only the edit forms with a single
field. This is so that autofocus is only applied when you would always
want to focus the first field and to avoid the on-screen keyboard on
mobile devices from hiding fields in multi-field forms on page load.